requirements.txt
aiohttp==3.5.4
aiohttp-devtools==0.13
aiohttp-session==2.7.0
alembic==1.0.10
asn1crypto==0.24.0
async-timeout==3.0.1
attrs==19.1.0
cffi==1.12.3
chardet==3.0.4
Click==7.0
cryptography==2.6.1
devtools==0.5
idna==2.8
idna-ssl==1.1.0
Mako==1.0.10
MarkupSafe==1.1.1
multidict==4.5.2
psycopg2-binary==2.8.2
pycparser==2.19
Pygments==2.4.2
python-dateutil==2.8.0
python-dotenv==0.10.2
python-editor==1.0.4
six==1.12.0
SQLAlchemy==1.3.4
SQLAlchemy-Utils==0.33.11
typing-extensions==3.7.2
watchgod==0.4
yarl==1.3.0
